// Loopy-fe/src/types/mySteps.ts
export type MyPageStep =
  | "my"
  | "setting"
  | "editProfile"
  | "manageAccount"
  | "withdraw"
  | "stampExchange"
  | "myChallenge"
  | "couponBox"
  | "stampHistory"
  | "review"
  | "filter"
  | "cafeNotice";

export type SettingStep = "settingMain" | "editProfile" | "manageAccount" | "withdraw";